+++ spec/spec_helper.rb
@@ -81,8 +81,10 @@ RSpec.configure do |c|
   c.disable_monkey_patching!
 
   # runtime options
-  c.raise_errors_for_deprecations!
-  c.color = true
+
+  # Temporary measure to prevent failures when run in a sub-build from other repos
+  # c.raise_errors_for_deprecations!
+
   c.include CommonHelpers
 
   c.expect_with :rspec do |expectations|